Improve project analysis (#2)
* Add `BriocheImportSpecifier` type

* Reorganize types in `specifier` module

* Temporarily add sync / async versions of `read_document` and `resolve`

The sync versions are specifically needed because the TypeScript
Language Service expects functions like `read_file` to be sync. This can
be resolved by reading all the files ahead-of-time, which can be done as
a follow-up

* Fix `find_project_root` when called with a file path

* Improve project analysis to find all imported modules

* Add `brioche analyze` command

This command mainly exists for testing

* Refactor out to add a separate `analyze::find_imports` function

* Add `runtime_specifiers_with_contents` function to iterate over runtime files

* Refactor `CompilerHost` to separate loading and reading documents

`CompilerHost` can now load a file, which loads it and all its imports
(recursively) into the list of documents. Reading a document now only
blocks on the `RwLock`, and never does I/O