Fix flaky CI by running Cargo tests with `--release`
Tests were failing with a "no space left on device" error. After doing a
quick sanity check of running `df -h /` with and without the `--release`
flag, it looks like debug tests eat > 20 GB of disk space, and release
tests use ~4 GB. Running `df -h /` at the start of a run shows that we
have about 20 GB of free space to play with at the time of writing.
Therefore, we basically need to run with `--release`, at least for the
time being.