Tweak `Project` to avoid storing the lockfile contents directly
This effectively undoes the change from
36ae4c3b322b43d50d2f686c559c1fbc610d5b51. To make it so a workspace
project can be exported to the registry then imported locally, the
`Project` needs to be the same whether or not it's put back into a
workspace directory. When moved outside its workspace, the workspace
dependencies should translate into lockfile dependencies, and both
projects need to have the same hash.

To handle this, the `Project` structure no longer separates the
`lockfile` field from the `internalDependencies` field: there's now a
single `dependencies` field that combines both sources.