Bump default timeouts
Although in the steady state 10 second timeout is reasonable, in
practice p100 was observed to be much higher than that (if, for example,
the machine you are supposed to talk to needs to cold boot).

Given that just _one_ timeout terminates the entire process, it feels
like we can benefit from being quiet a bit more conservative here for
the time being.